Simple "recent blocks" implementation (part 1)
Refactored code that generates and sorts visible blocks so that
the order of recent blocks can be arbitrary, and recent blocks
appear in their own category.

Maximum of 8 recent blocks are shown, with most recently used
at the top. This is not persisted between editor sessions, there
is a TODO item in the usage module to highlight this needs to
be done.

The buttons in block-list which currently show core/text and core/image
have not been modified to use the recently used blocks yet,
this is to come in the next iteration.